pyproject.toml
[project]
name = "np-chatbot"
version = "0.1.0"
description = "YouTube live chat bot daemon"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "pydantic-settings>=2.0",
    "python-jsonpath>=2.0.2",
    "structlog>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["np_chatbot"]
